refactor: centralize base url resolver
